feat(acp): move acp support from goose-cli to goose
