+++ crates/goose/Cargo.toml
@@ -97,10 +97,11 @@ tokio-stream = "0.1.17"
 tempfile = "3.15.0"
 dashmap = "6.1"
 ahash = "0.8"
-tokio-util = "0.7.15"
+tokio-util = { version = "0.7.15", features = ["compat"] }
 unicode-normalization = "0.1"
 
 oauth2 = "5.0.0"
+agent-client-protocol = "0.4.3"
 
 [target.'cfg(target_os = "windows")'.dependencies]
 winapi = { version = "0.3", features = ["wincred"] }
