+++ crates/goose/src/acp/mod.rs
@@ -1,0 +1,3 @@
+pub mod agent;
+
+pub use agent::run_acp_agent;
