+++ crates/goose/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod acp;
 pub mod agents;
 pub mod config;
 pub mod context_mgmt;
